requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic[email]
email-validator
pytest
pytest-asyncio
httpx
loguru>=0.7.0
pydantic-settings
psycopg2-binary
python-multipart